Moved InputStreamDesc struct to a separate file
