+++ src/AvTranscoder/transcoder/Transcoder.hpp
@@ -8,56 +8,14 @@
 #include <AvTranscoder/profile/ProfileLoader.hpp>
 #include <AvTranscoder/stat/ProcessStat.hpp>
 
-#include "StreamTranscoder.hpp"
+#include <AvTranscoder/transcoder/InputStreamDesc.hpp>
+#include <AvTranscoder/transcoder/StreamTranscoder.hpp>
 
 #include <string>
 #include <vector>
 
 namespace avtranscoder
 {
-
-/**
- * @brief Structure to describe the source data to extract.
- */
-struct InputStreamDesc
-{
-
-    InputStreamDesc(const std::string& filename, const size_t streamIndex, const std::vector<size_t>& channelIndexArray)
-        : _filename(filename)
-        , _streamIndex(streamIndex)
-        , _channelIndexArray(channelIndexArray)
-    {
-    }
-
-    InputStreamDesc(const std::string& filename, const size_t streamIndex, const size_t channelIndex)
-        : _filename(filename)
-        , _streamIndex(streamIndex)
-        , _channelIndexArray()
-    {
-        _channelIndexArray.push_back(channelIndex);
-    }
-
-    InputStreamDesc(const std::string& filename, const size_t streamIndex)
-        : _filename(filename)
-        , _streamIndex(streamIndex)
-        , _channelIndexArray()
-    {
-    }
-
-    /**
-     * @return If a demultiplexing step will be done to extract the expected data.
-     */
-    bool demultiplexing() const { return !_channelIndexArray.empty(); }
-
-public:
-    std::string _filename;                  ///< Source file path.
-    size_t _streamIndex;                    ///< Source stream to extract.
-    std::vector<size_t> _channelIndexArray; ///< List of source channels to extract from the stream
-};
-
-#ifndef SWIG
-AvExport std::ostream& operator<<(std::ostream& flux, const InputStreamDesc& inputStreamDesc);
-#endif
 
 /**
  * @brief Enum to set a policy of how we manage the process in case of several streams.
